# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Politics & War multiplayer strategy game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nations/me/cities:
    get:
      operationId: getMyCities
      tags: [cities]
      summary: List my cities
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/City"

    post:
      operationId: buildCity
      tags: [cities]
      summary: Build a new city (costs money)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CityInput"
      responses:
        "201":
          description: City built
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
        "400":
          description: Insufficient funds or max cities reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /nations/{nationId}/cities:
    get:
      operationId: getCities
      tags: [cities]
      summary: Get cities for a nation
      parameters:
        - name: nationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/City"
  /nations/me/cities/{cityId}/improve:
    post:
      operationId: improveCity
      tags: [cities]
      summary: Upgrade infrastructure/commerce/industry in a city
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CityImprovementInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
        "400":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── MILITARY ───────────────────────────────────────────────────────────────
...
  /market:
    get:
      operationId: listMarketOffers
      tags: [market]
      summary: List market offers
      parameters:
        - name: resource
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [buy, sell]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketOffer"

    post:
      operationId: createMarketOffer
      tags: [market]
      summary: Post a buy or sell offer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketOfferInput"
      responses:
        "201":
          description: Offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketOffer"
        "400":
          description: Insufficient resources or funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /market/{offerId}:
    delete:
      operationId: cancelMarketOffer
      tags: [market]
      summary: Cancel an offer you posted
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /market/{offerId}/buy:
    post:
      operationId: acceptMarketOffer
      tags: [market]
      summary: Accept / purchase a market offer
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketAcceptInput"
      responses:
        "200":
          description: Trade completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResult"
        "400":
          description: Insufficient funds or offer expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── DIPLOMACY / MESSAGES ───────────────────────────────────────────────────
  /messages:
    get:
      operationId: listMessages
      tags: [diplomacy]
      summary: List diplomatic messages (inbox + sent)
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [inbox, sent]
            default: inbox
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiplomaticMessage"

    post:
      operationId: sendMessage
      tags: [diplomacy]
      summary: Send a diplomatic message to a nation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiplomaticMessage"
  /messages/{messageId}:
    get:
      operationId: getMessage
      tags: [diplomacy]
      summary: Get a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiplomaticMessage"

  # ─── MAP ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ─── AUTH ────────────────────────────────────────────────────────────────
...
    Resources:
      type: object
      required: [money, food, coal, oil, iron, bauxite, lead, uranium, gasoline, steel, munitions, aluminum]
      properties:
        money:
          type: number
        food:
          type: number
        coal:
          type: number
        oil:
          type: number
        iron:
          type: number
        bauxite:
          type: number
        lead:
          type: number
        uranium:
          type: number
        gasoline:
          type: number
        steel:
          type: number
        munitions:
          type: number
        aluminum:
          type: number

    # ─── CITIES ──────────────────────────────────────────────────────────────
    City:
      type: object
      required: [id, nationId, name, infrastructure, land, commerce, industry, createdAt]
      properties:
        id:
          type: integer
        nationId:
          type: integer
        name:
          type: string
        infrastructure:
          type: number
        land:
          type: number
        commerce:
          type: integer
        industry:
          type: integer
        population:
          type: integer
        income:
          type: number
        createdAt:
          type: string
          format: date-time
    CityInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
    CityImprovementInput:
      type: object
      required: [type, amount]
      properties:
        type:
          type: string
          enum: [infrastructure, land, commerce, industry]
        amount:
          type: number

    # ─── MILITARY ────────────────────────────────────────────────────────────
...
    MarketOffer:
      type: object
      required: [id, nationId, nationName, offerType, resource, quantity, pricePerUnit, createdAt]
      properties:
        id:
          type: integer
        nationId:
          type: integer
        nationName:
          type: string
        offerType:
          type: string
          enum: [buy, sell]
        resource:
          type: string
        quantity:
          type: number
        pricePerUnit:
          type: number
        createdAt:
          type: string
          format: date-time
    MarketOfferInput:
      type: object
      required: [offerType, resource, quantity, pricePerUnit]
      properties:
        offerType:
          type: string
          enum: [buy, sell]
        resource:
          type: string
          enum: [food, coal, oil, iron, bauxite, lead, uranium, gasoline, steel, munitions, aluminum]
        quantity:
          type: number
        pricePerUnit:
          type: number
    MarketAcceptInput:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: number
    TradeResult:
      type: object
      required: [resource, quantity, totalCost]
      properties:
        resource:
          type: string
        quantity:
          type: number
        totalCost:
          type: number
        updatedResources:
          $ref: "#/components/schemas/Resources"

    # ─── DIPLOMACY ───────────────────────────────────────────────────────────
    DiplomaticMessage:
      type: object
      required: [id, fromNationId, fromNationName, toNationId, toNationName, subject, body, read, createdAt]
      properties:
        id:
          type: integer
        fromNationId:
          type: integer
        fromNationName:
          type: string
        toNationId:
          type: integer
        toNationName:
          type: string
        subject:
          type: string
        body:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [toNationId, subject, body]
      properties:
        toNationId:
          type: integer
        subject:
          type: string
          maxLength: 100
        body:
          type: string
          maxLength: 5000

    # ─── MAP ─────────────────────────────────────────────────────────────────
...